Add support for grouping.
Only very small changes were applied to the existing code.
Great care has been taken to integrate as seamlessly, as possible.
Changes are mainly added in "getFeature" and "getBody".

getBody returns the whole structure for grouping and a single row
for the "classic" case without grouping, which can be a little
confusing.

The optics of the group-header is determined by an entry added
to print.css and by XTemplate configuration:
              headerPrefix          : this.groupHeaderPrefix,
              headerPostfixSingular : this.groupHeaderPostfixSingular,
              headerPostfixPlural   : this.groupHeaderPostfixPlural,
              postfixWithParens     : true,